backend/requirement.txt
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
fastapi
uvicorn
streamlit
python-dotenv
pinecone 
langchain-pinecone 
langchain-groq
langchain-tavily
requests 
langchain-huggingface
pydantic[email]>=2.0
httpx
python-multipart
PyJWT>=2.8.0
email-validator
docx2txt 